mis cambios respaldo

+++ src/lib/wall.js
@@ -1,29 +1,58 @@
 import {
-  addPost, renderRealTime, deletePost, addLike, authUser, editPost,
+  addPost, renderRealTime, deletePost, addLike, authUser, editPost, onAuthStateChanged, auth,
 } from './index.js';
 
 export const renderWall = (navigateTo) => {
   const section = document.createElement('section');
   const template = `
-        <body class="wallbody">
+        <header class="wallbody">
           <img class="logo-wall" src="Images/logo_habitate_largo.png">
-          <section class="wallSectionInput" >
-          </section>
-          <section class="wallSection" >
-          </section>
-        </body>
-        <footer class="footer">
-            <button id="go-home" class="button-home"  > 
-                <img src="Images/home_habitate.png" class="image-home">
-            </button> 
-            <button id="go-profile" class="button-home"  > 
-                <img src="Images/perfil_habitate.png" class="image-home">
-            </button> 
-        </footer>
+        </header>
+        <div id="buttons-home-logout"
+          <button id="go-home" class="button-home"> 
+          <img src="Images/home_habitate.png" class="image-home">
+          </button> 
+       </div>
+        <section class="log-display">
+          <body>
+            <div class="wallSectionInput" >
+            </div>
+            <div class="wallSection" >
+            </div>
+          </body>
+          <footer class="footer">
+          </footer>
+        </section> 
       `;
 
   // Por ahora estoy creando aquí la sección de los post para poder empezar, posteriormente irá en una ventana pop up
   section.innerHTML = template;
+
+  auth.onAuthStateChanged((user) => {
+    console.log(user);
+    const displaySection = section.querySelector('.log-display');
+    if (user) {
+      displaySection.style.display = 'block';
+    } else {
+      displaySection.style.display = 'none';
+      const displayNoLog = document.createElement('div');
+      displayNoLog.className = 'log-out-display';
+      // displayNoLog.style.display = 'block';
+      const messageNoLog = document.createElement('p');
+      messageNoLog.innerText = 'Es necesario que inicies sesión para ver el contenido';
+      const buttonGoLogin = document.createElement('button');
+      buttonGoLogin.innerText = 'Inicia Sesión';
+      buttonGoLogin.id = 'go-login';
+      buttonGoLogin.className = 'button-second';
+      displayNoLog.append(messageNoLog, buttonGoLogin);
+      const header = section.querySelector('.wallbody');
+      header.append(displayNoLog);
+      buttonGoLogin.addEventListener('click', () => {
+        navigateTo('/login');
+      });
+    }
+  });
+
   const wallSectionInput = section.querySelector('.wallSectionInput');
   const wallSection = section.querySelector('.wallSection');
   const inputPost = document.createElement('input');
@@ -52,16 +81,19 @@ export const renderWall = (navigateTo) => {
   wallSectionInput.append(inputPost);
   wallSection.append(postType, buttonSendPost, postSection);
 
-  const originalEventListener = () => {
+  buttonSendPost.addEventListener('click', () => {
     const message = wallSectionInput.querySelector('#inPost');
     const postTypeSel = wallSection.querySelector('#select-type');
-    console.log('funciona el boton', message, postTypeSel);
-    const user = authUser();
-    const userID = user.uid;
-    addPost(message.value, postTypeSel.value, userID);
-    message.value = '';
-  };
-  buttonSendPost.addEventListener('click', originalEventListener);
+    if (message.value !== '') {
+      const user = authUser();
+      const userID = user.uid;
+      const userEmail = user.email;
+      addPost(message.value, postTypeSel.value, userID, userEmail);
+      message.value = '';
+    } else {
+      alert('El mensaje no puede estar vacío');
+    }
+  });
 
   renderRealTime((querySnapshot) => {
     postSection.textContent = '';
@@ -71,8 +103,15 @@ export const renderWall = (navigateTo) => {
       console.log(doc.data()); */
       const post = document.createElement('div');
       post.className = 'post-style';
+      post.id = 'post';
+      const userEmail = document.createElement('p');
+      userEmail.className = 'user-email';
+      userEmail.innerText = element.data().email;
+      const messageContainer = document.createElement('div');
+      messageContainer.className = 'messageContainer';
       const postMessage = document.createElement('p');
       postMessage.innerHTML = element.data().message;
+      postMessage.className = 'post-message-style';
       const btnEdit = document.createElement('button');
       btnEdit.id = 'button-edit';
       btnEdit.className = 'button-edit';
@@ -87,40 +126,70 @@ export const renderWall = (navigateTo) => {
       const apple = document.createElement('img');
       apple.src = 'Images/manzana_like.png';
       apple.className = 'img-like';
+      apple.style.opacity = 0.5;
       const counter = document.createElement('span');
       counter.innerText = element.data().likes.length;
       btnLike.append(apple);
-      post.append(postMessage, btnEdit, btnDelete, btnLike, counter);
+      messageContainer.append(postMessage);
+      post.append(userEmail, messageContainer, btnEdit, btnDelete, btnLike, counter);
       postSection.append(post);
       btnDelete.addEventListener('click', () => {
-        deletePost(docID);
+        const userID = authUser().uid;
+        const postUser = element.data().user;
+        console.log(userID, postUser);
+        if (postUser === userID) {
+          const confirmDelete = confirm('¿Seguro que deseas borrar este post?');
+          if (confirmDelete) {
+            deletePost(docID);
+          }
+        } else {
+          alert('Solo el autor original puede eliminar el post');
+        }
       });
       btnLike.addEventListener('click', () => {
         const user = authUser();
-        addLike(docID, user.uid);
+        addLike(docID, user.uid, apple);
+        apple.style.opacity = 1;
       });
       btnEdit.addEventListener('click', () => {
-        wallSectionInput.querySelector('#inPost').value = element.data().message;
-        buttonSendPost.textContent = 'Guardar';
-        buttonSendPost.removeEventListener('click', originalEventListener);
-        const buttonEditPost = () => {
-          const message = wallSectionInput.querySelector('#inPost');
-          const postTypeSel = wallSection.querySelector('#select-type');
-          editPost(docID, message.value, postTypeSel.value);
-          message.value = '';
-          buttonSendPost.removeEventListener('click', buttonEditPost);
-          buttonSendPost.addEventListener('click', originalEventListener);
-          buttonSendPost.textContent = 'Publicar';
-        };
-        buttonSendPost.addEventListener('click', buttonEditPost);
+        const userID = authUser().uid;
+        const postUser = element.data().user;
+        if (postUser === userID) {
+          const createInput = () => {
+            const inputEdit = document.createElement('input');
+            inputEdit.type = 'text';
+            inputEdit.className = 'input-edit';
+            inputEdit.id = 'input-edit';
+            inputEdit.value = element.data().message;
+            messageContainer.insertBefore(inputEdit, postMessage);
+            messageContainer.removeChild(postMessage);
+          };
+          const buttonEditPost = () => {
+            const inputEdit = wallSection.querySelector('#input-edit');
+            inputEdit.className = 'input-edit';
+            editPost(docID, inputEdit.value);
+            postMessage.innerHTML = element.data().message;
+            messageContainer.insertBefore(postMessage, inputEdit);
+            messageContainer.removeChild(inputEdit);
+            btnEdit.textContent = 'Editar';
+          };
+          if (btnEdit.textContent === 'Editar') {
+            createInput();
+            btnEdit.textContent = 'Guardar';
+            btnEdit.removeEventListener('click', buttonEditPost);
+            btnEdit.addEventListener('click', createInput);
+          } else {
+            buttonEditPost();
+          }
+        } else {
+          alert('Solo el autor original puede editar el post');
+        }
       });
     });
-  });
-
-  // Agrega el evento para direccionar al home de nuevo
-  const buttonHome = section.querySelector('#go-home');
-  buttonHome.addEventListener('click', () => {
-    navigateTo('/home');
+    const buttonHome = section.querySelector('#go-home');
+    buttonHome.addEventListener('click', () => {
+      navigateTo('/home');
+    });
   });
   return section;
 };
